requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.3.0=pyhd8ed1ab_0
aiohttp=3.8.1=py38h0a891b7_1
aiosignal=1.3.1=pyhd8ed1ab_0
async-timeout=4.0.2=py38h06a4308_0
attrs=22.1.0=pyh71513ae_1
blas=1.0=mkl
blinker=1.5=pyhd8ed1ab_0
brotlipy=0.7.0=py38h0a891b7_1004
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f98852_0
ca-certificates=2022.12.7=ha878542_0
cachetools=5.2.0=pyhd8ed1ab_0
certifi=2022.12.7=pyhd8ed1ab_0
cffi=1.15.1=py38h74dc2b5_0
charset-normalizer=2.1.1=pyhd8ed1ab_0
click=8.1.3=unix_pyhd8ed1ab_2
cloudpickle=2.2.0=pypi_0
contourpy=1.0.5=pypi_0
cryptography=3.4.8=py38h3e25421_1
cudatoolkit=11.1.74=h6bb024c_0
cycler=0.11.0=pypi_0
ffmpeg=4.2.2=h20bf706_0
fonttools=4.37.4=pypi_0
freetype=2.11.0=h70c0345_0
frozenlist=1.3.3=py38h5eee18b_0
giflib=5.2.1=h7b6447c_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth=2.15.0=pyh1a96a4e_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
grpcio=1.51.1=pypi_0
gym=0.26.2=pypi_0
gym-notices=0.0.8=pypi_0
idna=3.4=pyhd8ed1ab_0
importlib-metadata=5.1.0=pyha770c72_0
intel-openmp=2021.4.0=h06a4308_3561
jpeg=9b=h024ee3a_2
kiwisolver=1.4.4=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libidn2=2.3.2=h7f8727e_0
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.15.8=h780b84a_1
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.1.0=h2733197_1
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libvpx=1.7.0=h439df22_0
libwebp=1.2.0=h89dd481_0
lz4-c=1.9.3=h295c915_1
markdown=3.4.1=pyhd8ed1ab_0
markupsafe=2.1.1=py38h0a891b7_1
matplotlib=3.6.2=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py38h7f8727e_0
mkl_fft=1.3.1=py38hd3c417c_0
mkl_random=1.2.2=py38h51133e4_0
multidict=6.0.2=py38h0a891b7_1
ncurses=6.3=h5eee18b_3
nettle=3.7.3=hbbd107a_1
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
numpy=1.23.1=py38h6c91a56_0
numpy-base=1.23.1=py38ha15fc14_0
oauthlib=3.2.2=pyhd8ed1ab_0
opencv-python=4.6.0.66=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1s=h7f8727e_0
packaging=21.3=pypi_0
pandas=1.5.2=pypi_0
pillow=9.2.0=py38hace64e9_1
pip=22.1.2=py38h06a4308_0
protobuf=3.20.1=pypi_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pyjwt=2.6.0=pyhd8ed1ab_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyparsing=3.0.9=pypi_0
pysocks=1.7.1=pyha2e5f31_6
python=3.8.13=h12debd9_0
python-dateutil=2.8.2=pypi_0
python_abi=3.8=2_cp38
pytorch=1.8.2=py3.8_cuda11.1_cudnn8.0.5_0
pytz=2022.6=pypi_0
pyu2f=0.1.5=pyhd8ed1ab_0
readline=8.1.2=h7f8727e_1
requests=2.28.1=pyhd8ed1ab_1
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.9=pyhd8ed1ab_0
scipy=1.10.0=pypi_0
seaborn=0.12.1=pypi_0
setuptools=61.2.0=py38h06a4308_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.39.2=h5082296_0
tensorboard=2.11.0=pyhd8ed1ab_0
tensorboard-data-server=0.6.1=py38h52d8a92_0
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
tensorboardx=2.5.1=pypi_0
tk=8.6.12=h1ccaba5_0
torch-tb-profiler=0.4.0=pypi_0
torchaudio=0.8.2=py38
torchvision=0.9.2=py38_cu111
tqdm=4.64.1=pypi_0
typing_extensions=4.3.0=py38h06a4308_0
urllib3=1.26.13=pyhd8ed1ab_0
werkzeug=2.2.2=pyhd8ed1ab_0
wheel=0.37.1=pyhd3eb1b0_0
x264=1!157.20191217=h7b6447c_0
xz=5.2.5=h7f8727e_1
yarl=1.7.2=py38h0a891b7_2
zipp=3.11.0=pyhd8ed1ab_0
zlib=1.2.12=h7f8727e_2
zstd=1.4.9=haebb681_0
